requirements.txt
asgiref==3.8.1
contourpy==1.2.0
crispy-bootstrap4==2025.6
cycler==0.12.1
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.9.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
fonttools==4.44.3
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
kiwisolver==1.4.5
matplotlib==3.8.2
mysql-connector-python==8.0.31
mysqlclient==2.2.7
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
platformdirs==4.3.8
pygame==2.5.2
PyJWT==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==311
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2023.3
virtualenv==20.31.2
virtualenvwrapper-win==1.2.7
